Revert the IP to FQDN change and instead let the user decide.
This adds two new environment variables to foreman_server.sh where the
user can decide if they want a FQDN for the public and private interfaces.
If they want to use SSL, particularly with FreeIPA integration, they'll
need to use FQDNs.

If both are set then the IP wins. If neither is set it complains.

This also fixes a bug I introduced in the original IP to FQDN patch
where auto_assign_floating_ip was changed to auto_assign_fqdn.