BZ #1172315 - add support for keystone::ldap
https://bugzilla.redhat.com/show_bug.cgi?id=1172315
Added a lot of keystone LDAP identity backend class parameters
to quickstack::pacemaker::keystone.  The primary one is
keystone_identity_backend.  By default this is set to 'sql'.
If 'ldap' is used, this allows the use of the ldap related class
parameters, which are then in turn passed to keystone::ldap.
The keystone::ldap class parameters require true puppet undef values
instead of empty strings, so the function empty_to_undef was added
which will just pass through the values, converting empty strings to
the puppet undef value.

Note: This depends on
https://review.openstack.org/#/c/150041/
for juno and
https://review.openstack.org/#/c/133601/
for master