BZ #1180158: Ceilometer coordination resilience with redis-sentinel
https://bugzilla.redhat.com/show_bug.cgi?id=1180158

We use redis-sentinel to monitor the redis instances running on
controller nodes, and to detect when mastership needs to be failed
over to a slave.

The sentinel cluster has a default quorum of 2, so can survive the
loss of one controller on a minimally sized HA controller plane.

The ceilometer backend_url contains a list of fallback sentinels,
similar in style to the mongodb connect URL, except that the additional
addresses are specified as query parameters.