BZ #1190990 - Puppet restarts neutron server every run.
https://bugzilla.redhat.com/show_bug.cgi?id=1190990

We previously had to run db-manage in an exec due to no support in
puppet-neutron.  Our exec was flawed because it did not have refrshonly=>true,
making the db-manage run each time and restarting the neutron server.
Puppet-neutron now has support for sync_db, with all needed ordering deps, so
use that instead.