BZ#1204881 Move clone_opts and operation_opts from instance attributes.
https://bugzilla.redhat.com/show_bug.cgi?id=1204881

Due to known issues with pcs in 7.0 clone_opts and operation_opts would
become instance attributes and thus not work as expected. The following
changes are made to ensure this does not happen :

`interleave=true` is now set as a clone option.
`start-delay=` is now set as a operation parameter.
`start timeout=` is also set as a operation parameter.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>
Co-Authored-By: John Ruemker <jruemker@redhat.com>