BZ #1204483 - start+stop neutron-server on one node, first run
https://bugzilla.redhat.com/show_bug.cgi?id=1204483

Previously, a race condition existed where multiple nodes starting
neutron-server at the same time (by pacemaker) could cause a DB error
such as the one reported in this BZ (in this case, multiple nodes
attempted to update the table ml2_vxlan_allocations and then found it
in an unexpected state).  To avoid the issue, neutron-server is
started and then stopped on the neutron-vip node, thereby taking care
of any db-initialization tasks (one of which may be inserting into the
ml2_vxlan_allocations table), before neutron-server is added as a
cloned resource.